+++ doom_wasm.c
@@ -1,72 +1,28 @@
-#include "doomgeneric.h"
-
 #include <stdbool.h>
 #include <stdio.h>
 #include <string.h>
-#include <unistd.h>
+
+#include "doomgeneric.h"
+#include "doom_wasm.h"
 
 /*
- * This file provides augmentations to the interface (i.e. imports and exports)
- * that is produced when Doom is compiled to a WebAssembly module.
+ * This file reconciles the WebAssembly interface (i.e. imports/exports)
+ * specified in `doom_wasm.h` with Doom's interface specified in
+ * `doomgeneric.h`.
+ *
+ * In particular this file is responsible for two things:
+ *  1) Implementing all functions exported via WebAssembly,
+ *      via the features provided by `doomgeneric.h`.
+ *  2) Implementing all functions needed by `doomgeneric.h`,
+ *      via the functions imported via WebAssembly.
  */
 
 // *****************************************************************************
-// *                             IMPORTED FUNCTIONS                            *
-// *****************************************************************************
-
-#define IMPORT_MODULE(moduleName) __attribute__((import_module(moduleName)))
-
-IMPORT_MODULE("loading") void onGameInit(int width, int height);
-IMPORT_MODULE("loading")
-void getWadsSizes(int *numberOfWads, size_t *numberOfTotalBytesInAllWads);
-IMPORT_MODULE("loading")
-void readDataForAllWads(unsigned char *wadDataDestination,
-                        int *byteLengthOfEachWad);
-
-IMPORT_MODULE("ui") void drawFrame(uint32_t *screenBuffer);
-IMPORT_MODULE("ui") void setWindowTitle(const char *title); // TODO: remove this
-
-IMPORT_MODULE("console") void info(const char *message, size_t length);
-IMPORT_MODULE("console") void error(const char *message, size_t length);
-
-IMPORT_MODULE("runtimeControl") void sleepMs(uint32_t ms);
-IMPORT_MODULE("runtimeControl") uint32_t getTicksMs();
-IMPORT_MODULE("runtimeControl") void onExit(int32_t exitCode);
-
-// Return the size of the associated save game data, 0 in the case that no save
-// data exists for this slot
-IMPORT_MODULE("gameSaving") size_t sizeOfSaveGame(int gameSaveId);
-// Return the number of bytes read
-IMPORT_MODULE("gameSaving")
-size_t readSaveGame(int gameSaveId, unsigned char *dataDestination);
-// Return the number of bytes written
-IMPORT_MODULE("gameSaving")
-size_t writeSaveGame(int gameSaveId, unsigned char *data, size_t length);
-
-// *****************************************************************************
-// *                             EXPORTED FUNCTIONS                            *
-// *****************************************************************************
-
-#define EXPORT __attribute__((visibility("default")))
-
-EXPORT void initGame();
-EXPORT void tickGame();
-EXPORT void reportKeyDown(uint8_t doomKey);
-EXPORT void reportKeyUp(uint8_t doomKey);
-
-// Declarations for an implementation of all wasi-snapshot-preview1 functions
-// that Doom depends upon. All related function names are prefixed with
-// `internal__`.
-EXPORT void internal__proc_exit(int32_t exitCode);
-EXPORT int32_t internal__fd_fdstat_get(int32_t fd, __wasi_fdstat_t *fdstat);
-EXPORT int32_t internal__fd_seek(int32_t fd, int64_t offset, int32_t whence,
-                                 int32_t offset_out_ptr);
-EXPORT int32_t internal__fd_write(int32_t fd, const __wasi_ciovec_t *iovs,
-                                  int32_t iovs_len, uint32_t *nwritten);
-EXPORT int32_t internal__fd_close(int32_t fd);
-
-// *****************************************************************************
-// *                           IMPLEMENTATION DETAILS                          *
+// *                                    (1)                                    *
+// *                                                                           *
+// *             Implement all functions exported via WebAssembly              *
+// *               via the features provided by `doomgeneric.h`                *
+// *                                                                           *
 // *****************************************************************************
 
 // A global cache of which keys are currently pressed down, according to what
@@ -74,10 +30,30 @@ EXPORT int32_t internal__fd_close(int32_t fd);
 // `reportKeyUp`.
 static bool isKeyPressed[UINT8_MAX] = {false};
 
-////////////////////////////////////////////////////////////////////////////////
-// Here we implement the doomgeneric interface via the functions imported via
-// WebAssembly imports.
-////////////////////////////////////////////////////////////////////////////////
+void initGame() {
+  // Provide 0 command line arguments to Doom.
+  // Any configuration knobs we end up exposing via WebAssembly that would
+  // usually be handled by command line arguments will instead be handled in a
+  // different way.
+  int argc = 0;
+  char *argv[] = {};
+
+  doomgeneric_Create(argc, argv);
+}
+
+void tickGame() { doomgeneric_Tick(); }
+
+void reportKeyDown(uint8_t doomKey) { isKeyPressed[doomKey] = true; }
+
+void reportKeyUp(uint8_t doomKey) { isKeyPressed[doomKey] = false; }
+
+// *****************************************************************************
+// *                                    (2)                                    *
+// *                                                                           *
+// *            Implement all functions needed by `doomgeneric.h`              *
+// *                via the functions imported via WebAssembly                 *
+// *                                                                           *
+// *****************************************************************************
 
 void DG_Init() { onGameInit(DOOMGENERIC_RESX, DOOMGENERIC_RESY); }
 
@@ -331,143 +307,4 @@ void DG_PCXScreenshotTaken(unsigned char *screenshotBytes,
   // is enabled via the  "-devparm" command line argument, which we don't
   // currently allow in this implementation (in fact, we don't allow the passing
   // of ANY command line arguments to Doom).
-}
-
-////////////////////////////////////////////////////////////////////////////////
-// Here we implement, via the doomgeneric interface, the WebAssembly exported
-// functions.
-////////////////////////////////////////////////////////////////////////////////
-
-void initGame() {
-  // Provide 0 command line arguments to Doom.
-  // Any configuration knobs we end up exposing via WebAssembly that would
-  // usually be handled by command line arguments will instead be handled in a
-  // different way.
-  int argc = 0;
-  char *argv[] = {};
-
-  doomgeneric_Create(argc, argv);
-}
-
-void tickGame() { doomgeneric_Tick(); }
-
-void reportKeyDown(uint8_t doomKey) { isKeyPressed[doomKey] = true; }
-
-void reportKeyUp(uint8_t doomKey) { isKeyPressed[doomKey] = false; }
-
-////////////////////////////////////////////////////////////////////////////////
-// Here we implement all needed parts of the wasi-snapshot-preview1 interface
-// via functions imported via WebAssembly imports.
-////////////////////////////////////////////////////////////////////////////////
-
-// Documention on WASI preview 1
-//    https://github.com/WebAssembly/WASI/blob/main/legacy/preview1/docs.md
-//
-// Many macros and struct definitions used here are taken from wasi/api.h
-//    https://github.com/WebAssembly/wasi-libc/blob/main/libc-bottom-half/headers/public/wasi/api.h
-//
-// which, apparently, is automaticaly included when a source file is compiled
-// via the wasi-sdk, as we're doing.
-
-void internal__proc_exit(int32_t exitCode) { onExit(exitCode); }
-
-int32_t internal__fd_fdstat_get(int32_t fd, __wasi_fdstat_t *fdstat) {
-
-  // The only file descriptors this wasi-snapshot-preview1 implementation
-  // supports are STDOUT_FILENO and STDERR_FILENO. Those file descriptors are
-  // character devices to which the user can write, and that's it.
-  if (fd == STDOUT_FILENO || fd == STDERR_FILENO) {
-    fdstat->fs_filetype = __WASI_FILETYPE_CHARACTER_DEVICE;
-    fdstat->fs_flags = 0; // no flags
-    fdstat->fs_rights_base = __WASI_RIGHTS_FD_WRITE;
-    fdstat->fs_rights_inheriting = 0; // no rights are inherited
-
-    return __WASI_ERRNO_SUCCESS; // SUCCESS
-  } else {
-    return __WASI_ERRNO_BADF;
-  }
-}
-
-int32_t internal__fd_seek(int32_t fd, int64_t offset, int32_t whence,
-                          int32_t offset_out_ptr) {
-  // Seeking is not allowed on any file descriptors we support.
-  return __WASI_ERRNO_NOTSUP;
-}
-
-// The imported functions `info` and `error` shall be called exactly once for
-// each line of output written to stdout and stderr, respectively. To support
-// that we need to buffer the output we receive until we encounter a newline
-// character.
-//
-// We'll just leverage buffers with a fixed size for now, one for stdout and one
-// for stderr. This will very likely meet our needs as it's unlikely that a
-// single line written out to either stdout or stderr (the only two file
-// descriptors we support) will be longer than the number of characters our
-// fixed-sized buffer supports.
-//
-// If we're in the unlikely situation that we're about to overflow the buffer
-// we'll just flush the current contents of the buffer, effectively inserting
-// our own newline to enforce that no line flushed to the user is longer than
-// the buffer size.
-#define LINE_BUFFER_BYTE_LENGTH 1024
-static char lineBuffer_info[LINE_BUFFER_BYTE_LENGTH] = {0};
-static size_t lengthOfCurrentLine_info = 0;
-static char lineBuffer_error[LINE_BUFFER_BYTE_LENGTH] = {0};
-static size_t lengthOfCurrentLine_error = 0;
-
-int32_t internal__fd_write(int32_t fd, const __wasi_ciovec_t *iovs,
-                           int32_t iovs_len, uint32_t *nwritten) {
-
-  void (*characterSink)(const char *message, size_t length) =
-      (fd == STDOUT_FILENO) ? info : error;
-  char *lineBuffer = (fd == STDOUT_FILENO) ? lineBuffer_info : lineBuffer_error;
-  size_t *lengthOfCurrentLine_ptr = (fd == STDOUT_FILENO)
-                                        ? &lengthOfCurrentLine_info
-                                        : &lengthOfCurrentLine_error;
-
-  if (fd == STDOUT_FILENO || fd == STDERR_FILENO) {
-    *nwritten = 0;
-
-    for (int i = 0; i < iovs_len; i++) {
-      size_t charsConsumed = 0;
-      while (charsConsumed < iovs[i].buf_len) {
-        char nextCharacter = (char)iovs[i].buf[charsConsumed];
-        // If the next character is a newline character then we'll flush the
-        // buffer, not including the newline, to the user
-        if (nextCharacter == '\n') {
-          characterSink(lineBuffer, *lengthOfCurrentLine_ptr);
-          *lengthOfCurrentLine_ptr = 0;
-        } else {
-          // Otherwise, copy the next character from the buffer to the line
-          // buffer
-
-          // If we're about to overflow the buffer then we'll flush the buffer
-          // to the user
-          if ((*lengthOfCurrentLine_ptr) + 1 >= LINE_BUFFER_BYTE_LENGTH) {
-            characterSink(lineBuffer, *lengthOfCurrentLine_ptr);
-            *lengthOfCurrentLine_ptr = 0;
-          }
-
-          lineBuffer[*lengthOfCurrentLine_ptr] = nextCharacter;
-          (*lengthOfCurrentLine_ptr)++;
-        }
-
-        charsConsumed++;
-      }
-
-      *nwritten += iovs[i].buf_len;
-    }
-
-    return __WASI_ERRNO_SUCCESS;
-  } else {
-    return __WASI_ERRNO_BADF;
-  }
-}
-
-int32_t internal__fd_close(int32_t fd) {
-  // Only file descriptors this wasi-implementation supports are STDOUT_FILENO
-  // and STDERR_FILENO, which we don't expect the user to ever 'close', so this
-  // function should never be called, so we're just going to report that this
-  // closing a file descriptor is not supported.
-  return __WASI_ERRNO_NOTSUP;
 }
